LayoutTests/fast/dom/HTMLTableElement/insertRow-default-argument.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/tabular-data.html#htmltableelement">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Tests that HTMLTableElement.insertRow()'s default argument is -1.");

var table = document.createElement("table");
shouldBe("table.__proto__", "HTMLTableElement.prototype");

table.insertRow().innerHTML = "1"; // Should append.
table.insertRow().innerHTML = "2"; // Should append.
table.insertRow(-1).innerHTML = "3"; // Should append.
table.insertRow(0).innerHTML = "0"; // Should prepend.

var rows = table.rows;
shouldBe("rows.length", "4");
shouldBeEqualToString("rows[0].innerHTML", "0");
shouldBeEqualToString("rows[1].innerHTML", "1");
shouldBeEqualToString("rows[2].innerHTML", "2");
shouldBeEqualToString("rows[3].innerHTML", "3");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
